+++ src/geometry/primitives/Ray.cpp
@@ -4,7 +4,7 @@
 
 #include "Ray.h"
 
-Ray::Ray(const glm::vec3& origin, const glm::vec3& axis)
+Ray::Ray(const glm::dvec3& origin, const glm::dvec3& axis)
     : origin(origin)
     , axis(axis)
 {
